Exporter: initial support for `databricks_catalog`, `databricks_schema` and `databricks_volume`
Initial support for exporting of `databricks_catalog`, `databricks_schema` and
`databricks_volume`.  Code currently emits not supported resources, such as,
`databricks_grants`, `databricks_connection`, etc. - support for them will be in the
subsequent PRs.

Also, includes a bit of refactoring of the repeatable code for emitting in the incremental
mode.

Tests will be added separately when we get support for more resources, such as grants, etc.